# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restomate API",
        "description": "API de réservation de tables de restaurant. Gérez restaurants, tables et réservations avec authentification Sanctum.",
        "contact": {
            "name": "Alex AGO",
            "email": "alexyessougnonago@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://restomate-api.test/api/v1",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/restaurants/{restaurantId}/tables": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "Liste les tables d'un restaurant",
                "operationId": "8a05cd7e6d301a2d2aa1d185bbdd1069",
                "parameters": [
                    {
                        "name": "restaurantId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tables récupérées"
                    },
                    "404": {
                        "description": "Restaurant introuvable"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tables"
                ],
                "summary": "Créer une table (Admin)",
                "operationId": "cc0cbf2435d20a1189cf5e25c77fdfec",
                "parameters": [
                    {
                        "name": "restaurantId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "capacity",
                                    "location"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Table VIP 1"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "terrasse",
                                        "enum": [
                                            "intérieur",
                                            "terrasse",
                                            "bar",
                                            "privé"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Table créée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/restaurants/{restaurantId}/tables/available": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "Tables disponibles selon date, heure et capacité",
                "description": "Retourne les tables disponibles pour un créneau donné",
                "operationId": "d3b2f9d830d8399ce40e487d609575f7",
                "parameters": [
                    {
                        "name": "restaurantId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-04-01"
                        }
                    },
                    {
                        "name": "start_time",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "19:30"
                        }
                    },
                    {
                        "name": "guests",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 4
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tables disponibles retournées"
                    },
                    "422": {
                        "description": "Hors des horaires du restaurant"
                    }
                }
            }
        },
        "/restaurants/{restaurantId}/tables/{tableId}": {
            "put": {
                "tags": [
                    "Tables"
                ],
                "summary": "Modifier une table (Admin)",
                "operationId": "8f49c7c713653686e72d90c5c70c1bff",
                "parameters": [
                    {
                        "name": "restaurantId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "tableId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Table A"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 6
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Table mise à jour"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tables"
                ],
                "summary": "Archiver une table (Admin)",
                "operationId": "36e22475f8d0fea6c1699b58d85d0be7",
                "parameters": [
                    {
                        "name": "restaurantId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "tableId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Table archivée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entre ton token Sanctum obtenu après login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}